backend/requirements.txt
Django==3.1.2
mysqlclient==2.0.1
djangorestframework==3.12.1
